+++ src/types/models/EmailTemplateContent.d.ts
@@ -16,26 +16,22 @@
 import { Resource } from '../resource';
 import { Client } from '../client';
 import { OptionalKnownProperties } from '../optional-known-properties-type';
-import { PolicyRuleActionsEnroll } from './PolicyRuleActionsEnroll';
-import { IdpPolicyRuleAction } from './IdpPolicyRuleAction';
-import { PasswordPolicyRuleAction } from './PasswordPolicyRuleAction';
-import { OktaSignOnPolicyRuleSignonActions } from './OktaSignOnPolicyRuleSignonActions';
 
-declare class PolicyRuleActions extends Resource {
+
+declare class EmailTemplateContent extends Resource {
   constructor(resourceJson: Record<string, unknown>, client: Client);
 
-  enroll: PolicyRuleActionsEnroll;
-  idp: IdpPolicyRuleAction;
-  passwordChange: PasswordPolicyRuleAction;
-  selfServicePasswordReset: PasswordPolicyRuleAction;
-  selfServiceUnlock: PasswordPolicyRuleAction;
-  signon: OktaSignOnPolicyRuleSignonActions;
+  readonly _links: {[name: string]: unknown};
+  body: string;
+  fromAddress: string;
+  fromName: string;
+  subject: string;
 
 }
 
-type PolicyRuleActionsOptions = OptionalKnownProperties<PolicyRuleActions>;
+type EmailTemplateContentOptions = OptionalKnownProperties<EmailTemplateContent>;
 
 export {
-  PolicyRuleActions,
-  PolicyRuleActionsOptions
+  EmailTemplateContent,
+  EmailTemplateContentOptions
 };