+++ src/types/models/EmailTemplateTranslation.d.ts
@@ -16,26 +16,19 @@
 import { Resource } from '../resource';
 import { Client } from '../client';
 import { OptionalKnownProperties } from '../optional-known-properties-type';
-import { PolicyRuleActionsEnroll } from './PolicyRuleActionsEnroll';
-import { IdpPolicyRuleAction } from './IdpPolicyRuleAction';
-import { PasswordPolicyRuleAction } from './PasswordPolicyRuleAction';
-import { OktaSignOnPolicyRuleSignonActions } from './OktaSignOnPolicyRuleSignonActions';
 
-declare class PolicyRuleActions extends Resource {
+
+declare class EmailTemplateTranslation extends Resource {
   constructor(resourceJson: Record<string, unknown>, client: Client);
 
-  enroll: PolicyRuleActionsEnroll;
-  idp: IdpPolicyRuleAction;
-  passwordChange: PasswordPolicyRuleAction;
-  selfServicePasswordReset: PasswordPolicyRuleAction;
-  selfServiceUnlock: PasswordPolicyRuleAction;
-  signon: OktaSignOnPolicyRuleSignonActions;
+  subject: string;
+  template: string;
 
 }
 
-type PolicyRuleActionsOptions = OptionalKnownProperties<PolicyRuleActions>;
+type EmailTemplateTranslationOptions = OptionalKnownProperties<EmailTemplateTranslation>;
 
 export {
-  PolicyRuleActions,
-  PolicyRuleActionsOptions
+  EmailTemplateTranslation,
+  EmailTemplateTranslationOptions
 };