+++ src/types/models/EmailTemplateType.d.ts
@@ -1,0 +1,64 @@
+/*!
+ * Copyright (c) 2017-present, Okta, Inc. and/or its affiliates. All rights reserved.
+ * The Okta software accompanied by this notice is provided pursuant to the Apache License, Version 2.0 (the "License.")
+ *
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *
+ * See the License for the specific language governing permissions and limitations under the License.
+ */
+
+
+/* THIS FILE IS AUTO-GENERATED - SEE CONTRIBUTOR DOCUMENTATION */
+
+declare enum EmailTemplateType {
+  EMAIL_WELCOME = 'email.welcome',
+  EMAIL_FORGOTPASSWORD = 'email.forgotPassword',
+  EMAIL_EMAILLINKRECOVERYRESETFACTOR = 'email.emailLinkRecoveryResetFactor',
+  EMAIL_EMAILLINKRECOVERYRESETPWDFACTOR = 'email.emailLinkRecoveryResetPwdFactor',
+  EMAIL_EMAILLINKRECOVERYADRESETFACTOR = 'email.emailLinkRecoveryAdResetFactor',
+  EMAIL_EMAILLINKRECOVERYADRESETPWDFACTOR = 'email.emailLinkRecoveryAdResetPwdFactor',
+  EMAIL_EMAILLINKRECOVERYLDAPRESETPWDFACTOR = 'email.emailLinkRecoveryLdapResetPwdFactor',
+  EMAIL_FORGOTPASSWORDDENIED = 'email.forgotPasswordDenied',
+  EMAIL_TEMPPASSWORD = 'email.tempPassword',
+  EMAIL_AD_WELCOME = 'email.ad.welcome',
+  EMAIL_AD_FORGOTPASSWORD = 'email.ad.forgotPassword',
+  EMAIL_AD_FORGOTPASSWORDRESET = 'email.ad.forgotPasswordReset',
+  EMAIL_SUNONE_WELCOME = 'email.sunone.welcome',
+  EMAIL_SUNONE_FORGOTPASSWORD = 'email.sunone.forgotPassword',
+  EMAIL_SUNONE_FORGOTPASSWORDDENIED = 'email.sunone.forgotPasswordDenied',
+  EMAIL_SELFSERVICEUNLOCK = 'email.selfServiceUnlock',
+  EMAIL_EMAILLINKRECOVERYUNLOCKFACTOR = 'email.emailLinkRecoveryUnlockFactor',
+  EMAIL_EMAILLINKRECOVERYADUNLOCKFACTOR = 'email.emailLinkRecoveryAdUnlockFactor',
+  EMAIL_EMAILLINKRECOVERYUNLOCKPWDFACTOR = 'email.emailLinkRecoveryUnlockPwdFactor',
+  EMAIL_EMAILLINKRECOVERYADUNLOCKPWDFACTOR = 'email.emailLinkRecoveryAdUnlockPwdFactor',
+  EMAIL_EMAILLINKRECOVERYLDAPUNLOCKPWDFACTOR = 'email.emailLinkRecoveryLdapUnlockPwdFactor',
+  EMAIL_AD_SELFSERVICEUNLOCK = 'email.ad.selfServiceUnlock',
+  EMAIL_PUSHVERIFYACTIVATION = 'email.pushVerifyActivation',
+  EMAIL_ACCOUNTLOCKOUT = 'email.accountLockout',
+  EMAIL_EMAILTRANSACTIONVERIFICATION = 'email.emailTransactionVerification',
+  EMAIL_EMAILLINKAUTHENTICATIONTRANSACTION = 'email.emailLinkAuthenticationTransaction',
+  EMAIL_EMAILACTIVATION = 'email.emailActivation',
+  EMAIL_SIGNINFROMNEWDEVICE = 'email.signInFromNewDevice',
+  EMAIL_REGISTRATIONACTIVATION = 'email.registrationActivation',
+  EMAIL_EMAILCHANGECONFIRMATION = 'email.emailChangeConfirmation',
+  EMAIL_EMAILNEWCHANGENOTIFICATION = 'email.emailNewChangeNotification',
+  EMAIL_EMAILNEWALREADYCHANGEDNOTIFICATION = 'email.emailNewAlreadyChangedNotification',
+  EMAIL_REGISTRATIONEMAILVERIFICATION = 'email.registrationEmailVerification',
+  EMAIL_EMAILLINKFACTORVERIFICATION = 'email.emailLinkFactorVerification',
+  EMAIL_SELFSERVICEUNLOCKONUNLOCKEDACCOUNT = 'email.selfServiceUnlockOnUnlockedAccount',
+  EMAIL_ENDUSERSCHEDULEDLIFECYCLESTATUSCHANGE = 'email.endUserScheduledLifecycleStatusChange',
+  EMAIL_ADMINUSERSLISTSCHEDULEDLIFECYCLESTATUSCHANGE = 'email.adminUsersListScheduledLifecycleStatusChange',
+  EMAIL_FACTORENROLLMENT = 'email.factorEnrollment',
+  EMAIL_FACTORRESET = 'email.factorReset',
+  EMAIL_AUTHENTICATORENROLLMENT = 'email.authenticatorEnrollment',
+  EMAIL_AUTHENTICATORRESET = 'email.authenticatorReset',
+  EMAIL_AUTOMATION = 'email.automation',
+  EMAIL_PASSWORDCHANGED = 'email.passwordChanged',
+}
+
+export {
+  EmailTemplateType
+};